Drop bogus method. Improve stratified sampling tests. Some refactoring of tests
