+++ tests/valuation/methods/test_gt_shapley.py
@@ -1,6 +1,7 @@
 from itertools import islice
 
 import numpy as np
+import pytest
 from numpy.testing import assert_allclose, assert_array_equal
 
 from pydvl.valuation.methods.gt_shapley import (
@@ -62,7 +63,8 @@ def test_gt_sampler(seed):
     assert_allclose(frequencies, expected, atol=0.02)
 
 
-def test_gt_setsize_strategy():
-    indices = np.arange(1, 10)
-    strategy = GroupTestingSampleSize(n_samples=1)
-    print(strategy.total_samples(len(indices)))
+@pytest.mark.parametrize("n_samples", [1, 27])
+def test_gt_setsize_strategy(n_samples):
+    indices = np.arange(10)
+    strategy = GroupTestingSampleSize(n_samples=n_samples)
+    np.testing.assert_allclose(n_samples, strategy.sample_sizes(len(indices)).sum())
